Rename / Delete PoseEstimatorSusbsystem
