+++ src/main/java/frc/robot/subsystems/PoseEstimatorSubsystem.java
@@ -1,3 +1,7 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
 package frc.robot.subsystems;
 
 import static edu.wpi.first.units.Units.Degree;
@@ -15,107 +19,25 @@ import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.geometry.Translation3d;
 import edu.wpi.first.math.kinematics.SwerveDriveOdometry;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
-import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
+import frc.robot.Constants.VisionConstants;
 
 public class PoseEstimatorSubsystem extends SubsystemBase {
 
-    private final SwerveDriveOdometry poseEstimator;
-    private final PhotonCamera[] photonCameras;
+  private CameraDetection FRONT_LEFT_DETECTION = null;
+  private CameraDetection FRONT_RIGHT_DETECTION = null;
+  private CameraDetection BACK_LEFT_DETECTION = null;
+  private CameraDetection BACK_RIGHT_DETECTION = null;
 
-    // Standard deviations for state and vision measurements
-    private static final double STATE_STD_DEV_POS = 0.03;
-    private static final double STATE_STD_DEV_HEADING = 0.02;
-    private static final double VISION_STD_DEV_POS = 0.9;
-    private static final double VISION_STD_DEV_HEADING = 0.9;
-
-    // Camera names from your PhotonVision config
-    private static final String[] CAMERA_NAMES = {
-            "Arducam_5", // index 0: front left (or front)
-            "Arducam_7", // index 1: right
-            "Arducam_2", // index 2: left (used for intake)
-            "Arducam_6" // index 3: front right (or front)
-    };
-
-    // Camera indices for side cameras used in intake operations
-    private static final int LEFT_CAMERA_INDEX = 1;
-    private static final int RIGHT_CAMERA_INDEX = 2;
-
-    /**
-     * Robot-to-Camera transforms.
-     * X = forward, Y = left, Z = up.
-     * Rotation3d(roll, pitch, yaw) in radians.
-     */
-    private final Transform3d[] robotToCams = {
-            // Front Left camera (index 0)
-            new Transform3d(
-                    new Translation3d(0.288, 0.1397, 0.119),
-                    new Rotation3d(0.0, -0.2617, 0.0)),
-
-            // Right camera (index 1)
-            new Transform3d(
-                    new Translation3d(-0.196, -0.26, 0.140),
-                    new Rotation3d(0, -1.16, -0.434) // Adjust yaw as needed
-            ),
-
-            // Left camera (index 2)
-            new Transform3d(
-                    new Translation3d(-0.196, 0.26, 0.140),
-                    new Rotation3d(0, -1.16, 0.434) // Adjust yaw as needed
-            ),
-            // Front Right Camera (index 3)
-            new Transform3d(
-                    new Translation3d(0.288, -0.1397, 0.119),
-                    new Rotation3d(0.0, -0.2617, 0.0)),
-    };
-
-    // A helper class to hold detection information per camera.
-    private static class CameraDetection {
-        int tagId;
-        double detectionTimestamp;
-        Rotation2d detectionHeading;
-        double distanceToTag;
-        double bearingToTagDeg;
-        double lateralOffsetToTag;
-        double xOffsetToTag;
-        double yOffsetToTag;
-        double tagOrientationErrorDeg;
-        int cameraIndex;
-
-        public CameraDetection(int tagId, double detectionTimestamp, Rotation2d detectionHeading,
-                double distanceToTag, double bearingToTagDeg, double lateralOffsetToTag,
-                double xOffsetToTag, double yOffsetToTag, double tagOrientationErrorDeg,
-                int cameraIndex) {
-            this.tagId = tagId;
-            this.detectionTimestamp = detectionTimestamp;
-            this.detectionHeading = detectionHeading;
-            this.distanceToTag = distanceToTag;
-            this.bearingToTagDeg = bearingToTagDeg;
-            this.lateralOffsetToTag = lateralOffsetToTag;
-            this.xOffsetToTag = xOffsetToTag;
-            this.yOffsetToTag = yOffsetToTag;
-            this.tagOrientationErrorDeg = tagOrientationErrorDeg;
-            this.cameraIndex = cameraIndex;
-        }
-    }
-
-    // Array storing detection info for each camera
-    private final CameraDetection[] cameraDetections = new CameraDetection[CAMERA_NAMES.length];
-
-    // Robot's current heading from gyro
-    private Rotation2d currentHeading = new Rotation2d();
-
+  private final SwerveDriveOdometry poseEstimator;
+  private Rotation2d currentHeading;
+  
+    /** Creates a new PoseEstimatorSubsystem. */
     public PoseEstimatorSubsystem(Pose2d initialPose) {
-        photonCameras = new PhotonCamera[CAMERA_NAMES.length];
-        for (int i = 0; i < CAMERA_NAMES.length; i++) {
-            photonCameras[i] = new PhotonCamera(CAMERA_NAMES[i]);
-            cameraDetections[i] = null;
-        }
-
-        poseEstimator = new SwerveDriveOdometry(
+      poseEstimator = new SwerveDriveOdometry(
                 Constants.DriveConstants.kDriveKinematics,
                 initialPose.getRotation(),
                 new SwerveModulePosition[] {
@@ -125,83 +47,176 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
                         new SwerveModulePosition()
                 });
     }
-
+  
     public void update(Rotation2d gyroRotation, SwerveModulePosition[] modulePositions) {
         // Store current heading for use in coordinate transformations
         this.currentHeading = gyroRotation;
 
         // Update pose estimator with odometry
         poseEstimator.update(gyroRotation, modulePositions);
-        double currentTime = Timer.getFPGATimestamp();
 
-        // Loop through each camera and update its detection info independently.
-        for (int camIdx = 0; camIdx < photonCameras.length; camIdx++) {
-            PhotonPipelineResult result = photonCameras[camIdx].getLatestResult();
-            if (!result.hasTargets()) {
-                cameraDetections[camIdx] = null;
-                SmartDashboard.putBoolean("Camera " + camIdx + " Has Detection", false);
-                continue;
-            }
+        /*
+         * FRONT LEFT CAMERA
+         */
 
-            PhotonTrackedTarget bestTarget = result.getBestTarget();
-            if (bestTarget == null) {
-                cameraDetections[camIdx] = null;
-                SmartDashboard.putBoolean("Camera " + camIdx + " Has Detection", false);
-                continue;
-            }
+        PhotonPipelineResult FL_result = VisionConstants.FRONT_LEFT_CAMERA.getLatestResult();
+        PhotonTrackedTarget FL_bestTarget = FL_result.getBestTarget();
+        Transform3d FL_cameraToTarget = FL_bestTarget.getBestCameraToTarget();
+        Transform3d FL_robotToTarget = VisionConstants.FRONT_LEFT_CAMERA_LOCATION.plus(FL_cameraToTarget);
+        Translation3d FL_translation = FL_robotToTarget.getTranslation();
+        Rotation3d FL_rotation = FL_robotToTarget.getRotation();
 
-            Transform3d cameraToTarget = bestTarget.getBestCameraToTarget();
-            if (cameraToTarget == null) {
-                cameraDetections[camIdx] = null;
-                SmartDashboard.putBoolean("Camera " + camIdx + " Has Detection", false);
-                continue;
-            }
+        // Create and store detection info for this camera.
+        CameraDetection FL_detection = new CameraDetection(
+          FL_bestTarget.getFiducialId(),
+          FL_result.getTimestampSeconds(),
+          gyroRotation,
+          FL_translation.getMeasureX().in(Meter),
+          FL_rotation.getMeasureZ().in(Degree),
+          FL_translation.getMeasureY().in(Meter),
+          FL_translation.getX(),
+          FL_translation.getY(),
+          FL_rotation.getMeasureZ().in(Degree),
+          VisionConstants.FRONT_LEFT_CAMERA);
 
-            // Get robot-relative transform for this camera.
-            Transform3d robotToTarget = robotToCams[camIdx].plus(cameraToTarget);
-            Translation3d translation = robotToTarget.getTranslation();
-            Rotation3d rotation = robotToTarget.getRotation();
+        FRONT_LEFT_DETECTION = FL_detection;
+        logCameraDetection(FRONT_LEFT_DETECTION);
 
-            // Create and store detection info for this camera.
-            CameraDetection detection = new CameraDetection(
-                    bestTarget.getFiducialId(),
-                    result.getTimestampSeconds(),
-                    gyroRotation,
-                    translation.getMeasureX().in(Meter),
-                    rotation.getMeasureZ().in(Degree),
-                    translation.getMeasureY().in(Meter),
-                    translation.getX(),
-                    translation.getY(),
-                    rotation.getMeasureZ().in(Degree),
-                    camIdx);
-            cameraDetections[camIdx] = detection;
+        /*
+         * FRONT RIGHT CAMERA
+         */
 
-            // Update dashboard with camera-specific info.
-            SmartDashboard.putBoolean("Camera " + camIdx + " Has Detection", true);
-            SmartDashboard.putNumber("Camera " + camIdx + " Last Detection Time", detection.detectionTimestamp);
-            SmartDashboard.putNumber("Camera " + camIdx + " Last Tag ID", detection.tagId);
-            SmartDashboard.putNumber("Camera " + camIdx + " DistanceToTag", detection.distanceToTag);
-            SmartDashboard.putNumber("Camera " + camIdx + " Tag Bearing (Deg)", detection.bearingToTagDeg);
-            SmartDashboard.putNumber("Camera " + camIdx + " Tag Orientation Error (Deg)",
-                    detection.tagOrientationErrorDeg);
-            SmartDashboard.putNumber("Camera " + camIdx + " LateralOffsetToTag", detection.lateralOffsetToTag);
-            SmartDashboard.putNumber("Camera " + camIdx + " xOffsetToTag", detection.xOffsetToTag);
-            SmartDashboard.putNumber("Camera " + camIdx + " yOffsetToTag", detection.yOffsetToTag);
-        }
+        PhotonPipelineResult FR_result = VisionConstants.FRONT_RIGHT_CAMERA.getLatestResult();
+        PhotonTrackedTarget FR_bestTarget = FR_result.getBestTarget();
+        Transform3d FR_cameraToTarget = FR_bestTarget.getBestCameraToTarget();
+        Transform3d FR_robotToTarget = VisionConstants.FRONT_RIGHT_CAMERA_LOCATION.plus(FR_cameraToTarget);
+        Translation3d FR_translation = FR_robotToTarget.getTranslation();
+        Rotation3d FR_rotation = FR_robotToTarget.getRotation();
+
+        // Create and store detection info for this camera.
+        CameraDetection FR_detection = new CameraDetection(
+          FR_bestTarget.getFiducialId(),
+          FR_result.getTimestampSeconds(),
+          gyroRotation,
+          FR_translation.getMeasureX().in(Meter),
+          FR_rotation.getMeasureZ().in(Degree),
+          FR_translation.getMeasureY().in(Meter),
+          FR_translation.getX(),
+          FR_translation.getY(),
+          FR_rotation.getMeasureZ().in(Degree),
+          VisionConstants.FRONT_RIGHT_CAMERA);
+
+        FRONT_RIGHT_DETECTION = FR_detection;
+        logCameraDetection(FRONT_RIGHT_DETECTION);
+        /*
+         * BACK LEFT CAMERA
+         */
+
+        PhotonPipelineResult BL_result = VisionConstants.BACK_LEFT_CAMERA.getLatestResult();
+        PhotonTrackedTarget BL_bestTarget = BL_result.getBestTarget();
+        Transform3d BL_cameraToTarget = BL_bestTarget.getBestCameraToTarget();
+        Transform3d BL_robotToTarget = VisionConstants.BACK_LEFT_CAMERA_LOCATION.plus(BL_cameraToTarget);
+        Translation3d BL_translation = BL_robotToTarget.getTranslation();
+        Rotation3d BL_rotation = BL_robotToTarget.getRotation();
+
+        // Create and store detection info for this camera.
+        CameraDetection BL_detection = new CameraDetection(
+          BL_bestTarget.getFiducialId(),
+          BL_result.getTimestampSeconds(),
+          gyroRotation,
+          BL_translation.getMeasureX().in(Meter),
+          BL_rotation.getMeasureZ().in(Degree),
+          BL_translation.getMeasureY().in(Meter),
+          BL_translation.getX(),
+          BL_translation.getY(),
+          BL_rotation.getMeasureZ().in(Degree),
+          VisionConstants.BACK_LEFT_CAMERA);
+
+        BACK_LEFT_DETECTION = BL_detection;
+        logCameraDetection(BACK_LEFT_DETECTION);
+
+        /*
+         * BACK RIGHT CAMERA
+         */
+
+        PhotonPipelineResult BR_result = VisionConstants.BACK_RIGHT_CAMERA.getLatestResult();
+        PhotonTrackedTarget BR_bestTarget = BR_result.getBestTarget();
+        Transform3d BR_cameraToTarget = BR_bestTarget.getBestCameraToTarget();
+        Transform3d BR_robotToTarget = VisionConstants.BACK_RIGHT_CAMERA_LOCATION.plus(BR_cameraToTarget);
+        Translation3d BR_translation = BR_robotToTarget.getTranslation();
+        Rotation3d BR_rotation = BR_robotToTarget.getRotation();
+
+        // Create and store detection info for this camera.
+        CameraDetection BR_detection = new CameraDetection(
+          BR_bestTarget.getFiducialId(),
+          BR_result.getTimestampSeconds(),
+          gyroRotation,
+          BR_translation.getMeasureX().in(Meter),
+          BR_rotation.getMeasureZ().in(Degree),
+          BR_translation.getMeasureY().in(Meter),
+          BR_translation.getX(),
+          BR_translation.getY(),
+          BR_rotation.getMeasureZ().in(Degree),
+          VisionConstants.BACK_RIGHT_CAMERA);
+
+        BACK_RIGHT_DETECTION = BR_detection;
+        logCameraDetection(BACK_RIGHT_DETECTION);
+    }
+
+    private void logCameraDetection(CameraDetection detection) {
+      String cameraName = detection.camera.getName();
+      // Update dashboard with camera-specific info.
+      SmartDashboard.putBoolean("Camera " + cameraName + " Has Detection", true);
+      SmartDashboard.putNumber("Camera " + cameraName + " Last Detection Time", detection.detectionTimestamp);
+      SmartDashboard.putNumber("Camera " + cameraName + " Last Tag ID", detection.tagId);
+      SmartDashboard.putNumber("Camera " + cameraName + " DistanceToTag", detection.distanceToTag);
+      SmartDashboard.putNumber("Camera " + cameraName + " Tag Bearing (Deg)", detection.bearingToTagDeg);
+      SmartDashboard.putNumber("Camera " + cameraName + " Tag Orientation Error (Deg)", detection.tagOrientationErrorDeg);
+      SmartDashboard.putNumber("Camera " + cameraName + " LateralOffsetToTag", detection.lateralOffsetToTag);
+      SmartDashboard.putNumber("Camera " + cameraName + " xOffsetToTag", detection.xOffsetToTag);
+      SmartDashboard.putNumber("Camera " + cameraName + " yOffsetToTag", detection.yOffsetToTag);
     }
 
     // Helper method to get the most recent detection across all cameras.
     private CameraDetection getMostRecentDetection() {
-        CameraDetection mostRecent = null;
-        for (CameraDetection detection : cameraDetections) {
-            if (detection != null) {
-                if (mostRecent == null || detection.detectionTimestamp > mostRecent.detectionTimestamp) {
-                    mostRecent = detection;
-                }
-            }
-        }
-        return mostRecent;
+      CameraDetection[] detections = {
+        FRONT_LEFT_DETECTION,
+        FRONT_RIGHT_DETECTION,
+        BACK_LEFT_DETECTION,
+        BACK_RIGHT_DETECTION
+      };
+      CameraDetection mostRecent = null;
+      for (CameraDetection detection : detections) {
+          if (detection != null) {
+              if (mostRecent == null || detection.detectionTimestamp > mostRecent.detectionTimestamp) {
+                  mostRecent = detection;
+              }
+          }
+      }
+      return mostRecent;
+  }
+
+  private CameraDetection getDetectionOf(PhotonCamera camera) {
+
+    if (camera.equals(VisionConstants.FRONT_LEFT_CAMERA)) {
+        return FRONT_LEFT_DETECTION;
     }
+
+    if (camera.equals(VisionConstants.FRONT_RIGHT_CAMERA)) {
+      return FRONT_RIGHT_DETECTION;
+    }
+
+    if (camera.equals(VisionConstants.BACK_LEFT_CAMERA)) {
+      return BACK_LEFT_DETECTION;
+    }
+
+    if (camera.equals(VisionConstants.BACK_RIGHT_CAMERA)) {
+      return BACK_RIGHT_DETECTION;
+    }
+
+    return null;
+        
+  }
 
     // Global getters (aggregated using the most recent detection)
     public int getLastDetectedTagId() {
@@ -209,9 +224,9 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
         return detection != null ? detection.tagId : -1;
     }
 
-    public int getLastDetectionCameraIndex() {
+    public PhotonCamera getLastDetectionCamera() {
         CameraDetection detection = getMostRecentDetection();
-        return detection != null ? detection.cameraIndex : -1;
+        return detection.camera;
     }
 
     public double getLastDetectionTimestamp() {
@@ -225,11 +240,8 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
      * @param cameraIndex The index of the camera.
      * @return The distance to the tag in meters, or NaN if no detection exists.
      */
-    public double getDistanceToTag(int cameraIndex) {
-        if (cameraIndex < 0 || cameraIndex >= CAMERA_NAMES.length) {
-            return Double.NaN;
-        }
-        CameraDetection detection = cameraDetections[cameraIndex];
+    public double getDistanceToTag(PhotonCamera camera) {
+        CameraDetection detection = getDetectionOf(camera);
         return detection != null ? detection.distanceToTag : Double.NaN;
     }
 
@@ -239,11 +251,8 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
      * @param cameraIndex The index of the camera.
      * @return The lateral offset in meters, or NaN if no detection exists.
      */
-    public double getLateralOffsetToTag(int cameraIndex) {
-        if (cameraIndex < 0 || cameraIndex >= CAMERA_NAMES.length) {
-            return Double.NaN;
-        }
-        CameraDetection detection = cameraDetections[cameraIndex];
+    public double getLateralOffsetToTag(PhotonCamera camera) {
+        CameraDetection detection = getDetectionOf(camera);
         return detection != null ? detection.lateralOffsetToTag : Double.NaN;
     }
 
@@ -265,11 +274,8 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
         return detection != null ? detection.tagOrientationErrorDeg : Double.NaN;
     }
 
-    public double getTagOrientationErrorDeg(int cameraIndex) {
-        if (cameraIndex < 0 || cameraIndex >= CAMERA_NAMES.length) {
-            return Double.NaN;
-        }
-        CameraDetection detection = cameraDetections[cameraIndex];
+    public double getTagOrientationErrorDeg(PhotonCamera camera) {
+        CameraDetection detection = getDetectionOf(camera);
         return detection != null ? detection.tagOrientationErrorDeg : Double.NaN;
     }
 
@@ -291,10 +297,6 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
         poseEstimator.resetPosition(gyroRotation, modulePositions, newPose);
     }
 
-    public boolean isRedAlliance() {
-        return DriverStation.getAlliance().equals(DriverStation.Alliance.Red);
-    }
-
     /**
      * Checks if any side camera (left or right) has detected a tag for intake
      * operations.
@@ -303,10 +305,10 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
      */
     public boolean hasSideCameraDetection() {
         double currentTime = Timer.getFPGATimestamp();
-        boolean hasLeftDetection = cameraDetections[LEFT_CAMERA_INDEX] != null &&
-                (currentTime - cameraDetections[LEFT_CAMERA_INDEX].detectionTimestamp) < 1.0;
-        boolean hasRightDetection = cameraDetections[RIGHT_CAMERA_INDEX] != null &&
-                (currentTime - cameraDetections[RIGHT_CAMERA_INDEX].detectionTimestamp) < 1.0;
+        boolean hasLeftDetection = BACK_LEFT_DETECTION != null &&
+                (currentTime - BACK_LEFT_DETECTION.detectionTimestamp) < 1.0;
+        boolean hasRightDetection = BACK_RIGHT_DETECTION != null &&
+                (currentTime - BACK_RIGHT_DETECTION.detectionTimestamp) < 1.0;
         return hasLeftDetection || hasRightDetection;
     }
 
@@ -317,10 +319,10 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
      */
     public boolean hasFrontCameraDetection() {
         double currentTime = Timer.getFPGATimestamp();
-        boolean hasLeftDetection = cameraDetections[0] != null &&
-                (currentTime - cameraDetections[0].detectionTimestamp) < 1.0;
-        boolean hasRightDetection = cameraDetections[3] != null &&
-                (currentTime - cameraDetections[3].detectionTimestamp) < 1.0;
+        boolean hasLeftDetection = FRONT_LEFT_DETECTION != null &&
+                (currentTime - FRONT_LEFT_DETECTION.detectionTimestamp) < 1.0;
+        boolean hasRightDetection = FRONT_RIGHT_DETECTION != null &&
+                (currentTime - FRONT_RIGHT_DETECTION.detectionTimestamp) < 1.0;
         return hasLeftDetection || hasRightDetection;
     }
 
@@ -330,34 +332,31 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
      * @return The index of the side camera with the most recent detection, or -1 if
      *         no detection.
      */
-    public int getLastSideCameraDetection() {
-        CameraDetection leftDetection = cameraDetections[LEFT_CAMERA_INDEX];
-        CameraDetection rightDetection = cameraDetections[RIGHT_CAMERA_INDEX];
+    public CameraDetection getLastSideCameraDetection() {
+        CameraDetection leftDetection = BACK_LEFT_DETECTION;
+        CameraDetection rightDetection = BACK_RIGHT_DETECTION;
 
         if (leftDetection == null && rightDetection == null) {
-            return -1;
+            return null;
         }
 
         if (leftDetection != null && rightDetection != null) {
-            return (leftDetection.detectionTimestamp > rightDetection.detectionTimestamp) ? LEFT_CAMERA_INDEX
-                    : RIGHT_CAMERA_INDEX;
+            return (leftDetection.detectionTimestamp > rightDetection.detectionTimestamp) ? BACK_LEFT_DETECTION
+                    : BACK_RIGHT_DETECTION;
         }
 
-        return leftDetection != null ? LEFT_CAMERA_INDEX : RIGHT_CAMERA_INDEX;
+        return leftDetection != null ? BACK_LEFT_DETECTION : BACK_RIGHT_DETECTION;
     }
 
     /**
      * Checks whether a specific camera has detected a tag recently.
      * 
-     * @param cameraIndex The index of the camera to check.
+     * @param camera The camera to check.
      * @return true if the camera has a detection within the last 0.6 seconds.
      */
-    public boolean hasCameraDetectedTag(int cameraIndex) {
-        if (cameraIndex < 0 || cameraIndex >= CAMERA_NAMES.length) {
-            return false;
-        }
+    public boolean hasCameraDetectedTag(PhotonCamera camera) {
         double currentTime = Timer.getFPGATimestamp();
-        CameraDetection detection = cameraDetections[cameraIndex];
+        CameraDetection detection = getDetectionOf(camera);
         return detection != null && (currentTime - detection.detectionTimestamp) < 0.6;
     }
 
@@ -367,11 +366,8 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
      * @param cameraIndex The index of the camera.
      * @return The timestamp of the last detection, or -1 if none.
      */
-    public double getLastCameraDetectionTimestamp(int cameraIndex) {
-        if (cameraIndex < 0 || cameraIndex >= CAMERA_NAMES.length) {
-            return -1.0;
-        }
-        CameraDetection detection = cameraDetections[cameraIndex];
+    public double getLastCameraDetectionTimestamp(PhotonCamera camera) {
+        CameraDetection detection = getDetectionOf(camera);
         return detection != null ? detection.detectionTimestamp : -1.0;
     }
 
@@ -381,25 +377,9 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
      * @param cameraIndex The index of the camera.
      * @return The tag ID, or -1 if none.
      */
-    public int getLastTagDetectedByCamera(int cameraIndex) {
-        if (cameraIndex < 0 || cameraIndex >= CAMERA_NAMES.length) {
-            return -1;
-        }
-        CameraDetection detection = cameraDetections[cameraIndex];
+    public int getLastTagDetectedByCamera(PhotonCamera camera) {
+        CameraDetection detection = getDetectionOf(camera);
         return detection != null ? detection.tagId : -1;
-    }
-
-    /**
-     * Gets the name of the camera at the specified index.
-     * 
-     * @param cameraIndex The index of the camera.
-     * @return The camera name, or null if invalid.
-     */
-    public String getCameraName(int cameraIndex) {
-        if (cameraIndex < 0 || cameraIndex >= CAMERA_NAMES.length) {
-            return null;
-        }
-        return CAMERA_NAMES[cameraIndex];
     }
 
     /**
@@ -453,4 +433,36 @@ public class PoseEstimatorSubsystem extends SubsystemBase {
         degrees = degrees % 360;
         return degrees < 0 ? degrees + 360 : degrees;
     }
+
+
+
+  // A helper class to hold detection information per camera.
+  private static class CameraDetection {
+      int tagId;
+      double detectionTimestamp;
+      Rotation2d detectionHeading;
+      double distanceToTag;
+      double bearingToTagDeg;
+      double lateralOffsetToTag;
+      double xOffsetToTag;
+      double yOffsetToTag;
+      double tagOrientationErrorDeg;
+      PhotonCamera camera;
+
+      public CameraDetection(int tagId, double detectionTimestamp, Rotation2d detectionHeading,
+              double distanceToTag, double bearingToTagDeg, double lateralOffsetToTag,
+              double xOffsetToTag, double yOffsetToTag, double tagOrientationErrorDeg,
+              PhotonCamera camera) {
+          this.tagId = tagId;
+          this.detectionTimestamp = detectionTimestamp;
+          this.detectionHeading = detectionHeading;
+          this.distanceToTag = distanceToTag;
+          this.bearingToTagDeg = bearingToTagDeg;
+          this.lateralOffsetToTag = lateralOffsetToTag;
+          this.xOffsetToTag = xOffsetToTag;
+          this.yOffsetToTag = yOffsetToTag;
+          this.tagOrientationErrorDeg = tagOrientationErrorDeg;
+          this.camera = camera;
+      }
+  }
 }
